vault main() cleanup
