+++ e2e/hyperlane/config/config.json
@@ -14,9 +14,9 @@
       "displayName": "Ethereum",
       "domainId": 1,
       "gasCurrencyCoinGeckoId": "ethereum",
-      "interchainGasPaymaster": "0x5FC8d32690cc91D4c39d9d3abcBD16989F875707",
-      "mailbox": "0xCf7Ed3AccA5a467e9e704C703E8D87F634fB0Fc9",
-      "merkleTreeHook": "0xDc64a140Aa3E981100a9becA4E685f962f0cF6C9",
+      "interchainGasPaymaster": "0x0165878A594ca255338adfa4d48449f69242Eb8F",
+      "mailbox": "0xDc64a140Aa3E981100a9becA4E685f962f0cF6C9",
+      "merkleTreeHook": "0x5FC8d32690cc91D4c39d9d3abcBD16989F875707",
       "name": "ethereum",
       "nativeToken": {
         "decimals": 18,
@@ -30,7 +30,7 @@
         }
       ],
       "technicalStack": "other",
-      "validatorAnnounce": "0xa513E6E4b8f2a923D98304ec87F64353C4D5C853"
+      "validatorAnnounce": "0x2279B7A0a67DB372996a5FaB50D91eAA73d2eBe6"
     },
     "neutron": {
       "bech32Prefix": "neutron",
@@ -58,10 +58,10 @@
           "http": "http://localneutron-1-val-0-neutronic:9090"
         }
       ],
-      "interchainGasPaymaster": "0x53759cc25ea77a1b7e348d4643a0caf86da6b0a06925eec2d9a3f2f3548dba3e",
+      "interchainGasPaymaster": "0x745e1c5172c3a3b453b6c16d95b5617f657a019d9f1281ca99372fcc8e6d361a",
       "isTestnet": false,
-      "mailbox": "0x7398fec2b158bb621d05a4b5eb9c9bca1cb72c7990ad8fdd1e75ae76af862445",
-      "merkleTreeHook": "0xe1ce9317242a22818a906a952a1ced1c26267c5c56ba1f3470fd4b48a6c23172",
+      "mailbox": "0xc735e23974a17512683549a1d29df3fcc8b5c37d50f9b137f61b3096c5345db0",
+      "merkleTreeHook": "0x8c39869ac9f855514bac8e43c5b00ff5caafd843d044f391a7692c2d407bc6c7",
       "name": "neutron",
       "nativeToken": {
         "decimals": 6,
@@ -87,7 +87,7 @@
       },
       "slip44": 118,
       "technicalStack": "other",
-      "validatorAnnounce": "0xc3b79277b7bc829496eea2f40022371bdc12749298e23541bba179c42389e78e"
+      "validatorAnnounce": "0x5d26fec10e8d8f33ff9a5fa964263226d63a5c27a365a1b24d0982a51a6c6f31"
     }
   },
   "defaultRpcConsensusType": "fallback"
